table definitions to separate package
