+++ src/main/java/nl/knaw/dans/migration/DdVerifyMigrationApplication.java
@@ -26,9 +26,9 @@ import io.dropwizard.setup.Environment;
 import nl.knaw.dans.migration.cli.LoadFromDataverseCommand;
 import nl.knaw.dans.migration.cli.LoadFromFedoraCommand;
 import nl.knaw.dans.migration.cli.LoadFromVaultCommand;
-import nl.knaw.dans.migration.core.ActualFile;
+import nl.knaw.dans.migration.core.tables.ActualFile;
 import nl.knaw.dans.migration.core.EasyFile;
-import nl.knaw.dans.migration.core.ExpectedFile;
+import nl.knaw.dans.migration.core.tables.ExpectedFile;
 
 public class DdVerifyMigrationApplication extends Application<DdVerifyMigrationConfiguration> {
