+++ src/main/java/nl/knaw/dans/migration/core/ExpectedLoader.java
@@ -15,6 +15,7 @@
  */
 package nl.knaw.dans.migration.core;
 
+import nl.knaw.dans.migration.core.tables.ExpectedFile;
 import nl.knaw.dans.migration.db.ExpectedFileDAO;
 import org.hibernate.exception.ConstraintViolationException;
 import org.slf4j.Logger;
