+++ src/main/java/nl/knaw/dans/migration/core/tables/ActualFile.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.migration.core;
+package nl.knaw.dans.migration.core.tables;
 
 import org.hsqldb.lib.StringUtil;
 import org.joda.time.DateTime;
